Update hlsdk dependency to support ReHLDS API Add C2S next client message handler

+++ deps/cssdk/dlls/effects.h
@@ -145,12 +145,12 @@ public:
 	Vector Center( void ) { return (GetStartPos() + GetEndPos()) * 0.5; }; // center point of beam
 
 	inline int  GetTexture( void ) { return pev->modelindex; }
-	inline int  GetWidth( void ) { return static_cast<int>(pev->scale); }
+	inline int  GetWidth( void ) { return pev->scale; }
 	inline int  GetNoise( void ) { return pev->body; }
 	// inline void GetColor( int r, int g, int b ) { pev->rendercolor.x = r; pev->rendercolor.y = g; pev->rendercolor.z = b; }
-	inline int  GetBrightness( void ) { return static_cast<int>(pev->renderamt); }
-	inline int  GetFrame( void ) { return static_cast<int>(pev->frame); }
-	inline int  GetScrollRate( void ) { return static_cast<int>(pev->animtime); }
+	inline int  GetBrightness( void ) { return pev->renderamt; }
+	inline int  GetFrame( void ) { return pev->frame; }
+	inline int  GetScrollRate( void ) { return pev->animtime; }
 
 	// Call after you change start/end positions
 	void		RelinkBeam( void );
